requirements.txt
pytz==2018.3
python-dateutil==2.7.2
psycopg2-binary==2.7.4
requests==2.18.4

django==2.0.3
djangorestframework==3.7.7
django-filter==1.1.0

django-cors-headers==2.1.0

coreapi
pygments
markdown
